requirements.txt
easydict==1.9
matplotlib==3.0.0
numpy==1.16.1
onnx==1.5.0
opencv-python==3.4.4.19
Pillow==6.2.0
protobuf==3.8.0
scipy==1.1.0
thop==0.0.31.post2001170342
torch==1.1.0
torchvision==0.3.0
tqdm==4.25.0
